Extraje logica de acceso a la productividad del back end a un hook personalizado.
