+++ client/src/hooks/data/index.js
@@ -4,3 +4,4 @@
 export * as empleadosData from "./Empleados";
 export * as cargosData from "./Cargos";
 export * as ciudadesData from "./Ciudades";
+export * as productividadData from "./Productividad";
